missing mailer unit test
