+++ tests/Unit/Domain.LinnApps.Tests/Domain.LinnApps.Tests.csproj
@@ -23,7 +23,6 @@
   <ItemGroup>
     <Folder Include="ForecastingServiceTests\" />
     <Folder Include="MrOrderBookReportServiceTests\" />
-    <Folder Include="MrOrderBookMailer\" />
     <Folder Include="PurchaseOrderDeliveryServiceTests\" />
     <Folder Include="WhatsInInspectionReportServiceTests\" />
     <Folder Include="OutstandingPoReqsReportServiceTests\" />
